[#13]: Refactoring the application framework - Move Gojira\Api\Configuration to Gojira\Framework\App\Configuration - Move Gojira\Api\Data to Gojira\Framework\Data - Move TableInterface to Gojira\Framework\App\Console - Added new available issues statuses to default option - Removed redundant 'use' statements - Fixed incorrect 'use' statements of few commands - Update .gitignore for /docs/ dir (ApiGen) - Added new ProgressBar provider as console component
