+++ src/Gojira/Framework/App/Configuration/Auth.php
@@ -6,14 +6,14 @@
  * file that was distributed with this source code.
  */
 
-namespace Gojira\Api\Configuration;
+namespace Gojira\Framework\App\Configuration;
 
-use Gojira\Api\Data\DataObject;
+use Gojira\Framework\Data\DataObject;
 
 /**
  * Base class to work with auth config items
  *
- * @package Gojira\Api\Configuration
+ * @package Gojira\Framework\App\Configuration
  * @author  Toan Nguyen <me@nntoan.com>
  */
 class Auth extends DataObject implements AuthInterface
