+++ src/Gojira/Framework/App/Configuration/AuthInterface.php
@@ -6,12 +6,12 @@
  * file that was distributed with this source code.
  */
 
-namespace Gojira\Api\Configuration;
+namespace Gojira\Framework\App\Configuration;
 
 /**
  * Interface for auth config items
  *
- * @package Gojira\Api\Configuration
+ * @package Gojira\Framework\App\Configuration
  * @author  Toan Nguyen <me@nntoan.com>
  */
 interface AuthInterface
